// src/app/model/tableStructDTO.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface TableStructDTO { 
    difference?: string;
    difftabid?: number;
    runid?: string;
    taskid?: string;
    tname?: string;
    tstructnewsys?: string;
    tstructoldsys?: string;
}